Remove value callback from `group`.
As Gavin points out, there's really no point to `group` taking a
callback to compute values.  Yes, the symmetry for keys and values
was nice and yes, it's extra work to `map` each of the groups' lists
afterwards and true, it might be nice to have each group be a Stream,
but it's a very straightforward extra step once you start processing
your result dict.